Begin solving race condition while serving cached files
I'm still not confident this is the right way to do; it's the first time I'm
using these __sync_*() builtins.  Have to read more about them before
merging this into master.

+++ lwan.h
@@ -132,7 +132,7 @@ struct lwan_key_value_t_ {
 struct lwan_response_t_ {
     strbuf_t *buffer;
     char *mime_type;
-    int content_length;
+    size_t content_length;
     lwan_key_value_t *headers;
 
     struct {
